requirements.txt
# requirements.txt
# API framework
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0

# Google APIs
google-generativeai>=0.3.0
google-api-python-client==2.103.0
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Twilio WhatsApp API
twilio>=7.0.0

# HTTP requests
aiohttp==3.8.6
requests==2.31.0

# Email handling
email-validator==2.0.0.post2

# Date handling
python-dateutil==2.8.2
pytz==2023.3

# For deployment (optional)
gunicorn==21.2.0

pytz